Create funsor.interpretations; remove interpretation(-)
