+++ funsor/__init__.py
@@ -32,6 +32,7 @@ from . import (  # minipyro,  # TODO: enable when minipyro is backend-agnostic
     einsum,
     gaussian,
     integrate,
+    interpretations,
     interpreter,
     joint,
     memoize,
@@ -74,6 +75,7 @@ __all__ = [
     "get_backend",
     "integrate",
     "interpretation",
+    "interpretations",
     "interpreter",
     "joint",
     "memoize",
