+++ funsor/adjoint.py
@@ -8,7 +8,6 @@ import numpy as np
 from funsor.cnf import Contraction, GaussianMixture, nullop
 from funsor.domains import Bint
 from funsor.gaussian import Gaussian, align_gaussian
-from funsor.interpreter import interpretation
 from funsor.ops import AssociativeOp
 from funsor.registry import KeyedRegistry
 from funsor.tensor import Tensor
@@ -46,7 +45,7 @@ class AdjointTape(object):
 
     def __call__(self, cls, *args):
         if cls in adjoint_ops:  # atomic op, don't trace internals
-            with interpretation(self._old_interpretation):
+            with self._old_interpretation:
                 result = cls(*args)
             self.tape.append((result, cls, args))
         else:
@@ -78,7 +77,7 @@ class AdjointTape(object):
                     continue
 
             # reverse the effects of alpha-renaming
-            with interpretation(reflect):
+            with reflect:
                 other_subs = tuple(
                     (name, to_funsor(name.split("__BOUND")[0], domain))
                     for name, domain in output.inputs.items()
