+++ funsor/einsum/__init__.py
@@ -5,7 +5,6 @@ from functools import reduce
 
 import funsor.ops as ops
 from funsor.cnf import Contraction
-from funsor.interpreter import interpretation
 from funsor.optimizer import apply_optimizer
 from funsor.sum_product import sum_product
 from funsor.terms import Funsor, Variable, lazy
@@ -124,6 +123,6 @@ def einsum(eqn, *terms, **kwargs):
         terms): dimensions in plates but not in outputs are product-reduced;
         dimensions in neither plates nor outputs are sum-reduced.
     """
-    with interpretation(lazy):
+    with lazy:
         naive_ast = naive_plated_einsum(eqn, *terms, **kwargs)
     return apply_optimizer(naive_ast)
