+++ test/test_integrate.py
@@ -9,9 +9,9 @@ import pytest
 from funsor import ops
 from funsor.domains import Bint
 from funsor.integrate import Integrate
-from funsor.interpreter import interpretation
+from funsor.interpretations import eager, lazy, moment_matching, normalize, reflect
 from funsor.montecarlo import MonteCarlo
-from funsor.terms import Variable, eager, lazy, moment_matching, normalize, reflect
+from funsor.terms import Variable
 from funsor.testing import assert_close, random_tensor
 
 
@@ -29,7 +29,7 @@ from funsor.testing import assert_close, random_tensor
 def test_integrate(interp):
     log_measure = random_tensor(OrderedDict([("i", Bint[2]), ("j", Bint[3])]))
     integrand = random_tensor(OrderedDict([("j", Bint[3]), ("k", Bint[4])]))
-    with interpretation(interp):
+    with interp:
         Integrate(log_measure, integrand, {"i", "j", "k"})
