+++ test/test_optimizer.py
@@ -13,7 +13,7 @@ from funsor.einsum import (
     naive_einsum,
     naive_plated_einsum,
 )
-from funsor.interpreter import interpretation, reinterpret
+from funsor.interpreter import reinterpret
 from funsor.optimizer import apply_optimizer
 from funsor.tensor import Tensor
 from funsor.terms import Variable, normalize, reflect
@@ -60,7 +60,7 @@ OPTIMIZED_EINSUM_EXAMPLES = [make_chain_einsum(t) for t in range(2, 50, 10)] + [
 def test_optimized_einsum(equation, backend, einsum_impl):
     inputs, outputs, sizes, operands, funsor_operands = make_einsum_example(equation)
     expected = pyro_einsum(equation, *operands, backend=backend)[0]
-    with interpretation(normalize):
+    with normalize:
         naive_ast = einsum_impl(equation, *funsor_operands, backend=backend)
     optimized_ast = apply_optimizer(naive_ast)
     actual = reinterpret(optimized_ast)  # eager by default
@@ -114,7 +114,7 @@ def test_nested_einsum(
         modulo_total=True
     )[0]
 
-    with interpretation(normalize):
+    with normalize:
         funsor_operands1 = [
             Categorical(
                 probs=Tensor(
@@ -126,12 +126,12 @@ def test_nested_einsum(
         ]
 
         output1_naive = einsum_impl(eqn1, *funsor_operands1, backend=backend1)
-        with interpretation(reflect):
+        with reflect:
             output1 = apply_optimizer(output1_naive) if optimize1 else output1_naive
         output2_naive = einsum_impl(
             outputs1[0] + "," + eqn2, *([output1] + funsor_operands2), backend=backend2
         )
-        with interpretation(reflect):
+        with reflect:
             output2 = apply_optimizer(output2_naive) if optimize2 else output2_naive
 
     actual1 = reinterpret(output1)
