+++ test/test_samplers.py
@@ -14,7 +14,6 @@ from funsor.delta import Delta
 from funsor.distribution import BACKEND_TO_DISTRIBUTIONS_BACKEND
 from funsor.domains import Bint, Real, Reals
 from funsor.integrate import Integrate
-from funsor.interpreter import interpretation
 from funsor.montecarlo import MonteCarlo
 from funsor.tensor import Tensor, align_tensors
 from funsor.terms import Variable
@@ -443,7 +442,7 @@ def test_lognormal_distribution(moment):
 
     log_measure = dist.LogNormal(loc, scale)(value="x")
     probe = Variable("x", Real) ** moment
-    with interpretation(MonteCarlo(particle=Bint[num_samples])):
+    with MonteCarlo(particle=Bint[num_samples]):
         with xfail_if_not_implemented():
             actual = Integrate(log_measure, probe, frozenset(["x"]))
