+++ test/test_sum_product.py
@@ -10,7 +10,6 @@ import pytest
 
 import funsor.ops as ops
 from funsor.domains import Bint, Real, Reals
-from funsor.interpreter import interpretation
 from funsor.optimizer import apply_optimizer
 from funsor.sum_product import (
     MarkovProduct,
@@ -126,7 +125,7 @@ def test_partial_sum_product(impl, sum_op, prod_op, inputs, plates, vars1, vars2
     expected = sum_product(sum_op, prod_op, factors, vars1 | vars2, frozenset(plates))
     assert_close(actual, expected)
 
-    with interpretation(lazy):
+    with lazy:
         unrolled_factors, unrolled_vars, remaining_plates = partial_unroll(
             factors, vars1 | vars2, plate_to_step
         )
@@ -188,7 +187,7 @@ def test_modified_partial_sum_product_0(sum_op, prod_op, vars1, vars2, x_dim, ti
     )
     actual = reduce(prod_op, factors2)
 
-    with interpretation(lazy):
+    with lazy:
         unrolled_factors, unrolled_vars, remaining_plates = partial_unroll(
             factors, vars1 | vars2, plate_to_step
         )
@@ -272,7 +271,7 @@ def test_modified_partial_sum_product_1(
     )
     actual = reduce(prod_op, factors2)
 
-    with interpretation(lazy):
+    with lazy:
         unrolled_factors, unrolled_vars, remaining_plates = partial_unroll(
             factors, vars1 | vars2, plate_to_step
         )
@@ -362,7 +361,7 @@ def test_modified_partial_sum_product_2(
     )
     actual = reduce(prod_op, factors2)
 
-    with interpretation(lazy):
+    with lazy:
         unrolled_factors, unrolled_vars, remaining_plates = partial_unroll(
             factors, vars1 | vars2, plate_to_step
         )
@@ -454,7 +453,7 @@ def test_modified_partial_sum_product_3(
     )
     actual = reduce(prod_op, factors2)
 
-    with interpretation(lazy):
+    with lazy:
         unrolled_factors, unrolled_vars, remaining_plates = partial_unroll(
             factors, vars1 | vars2, plate_to_step
         )
@@ -582,7 +581,7 @@ def test_modified_partial_sum_product_4(
     )
     actual = reduce(prod_op, factors2)
 
-    with interpretation(lazy):
+    with lazy:
         unrolled_factors, unrolled_vars, remaining_plates = partial_unroll(
             factors, vars1 | vars2, plate_to_step
         )
@@ -731,7 +730,7 @@ def test_modified_partial_sum_product_5(
     )
     actual = reduce(prod_op, factors2)
 
-    with interpretation(lazy):
+    with lazy:
         unrolled_factors, unrolled_vars, remaining_plates = partial_unroll(
             factors, vars1 | vars2, plate_to_step
         )
@@ -860,7 +859,7 @@ def test_modified_partial_sum_product_6(
     )
     actual = reduce(prod_op, factors2)
 
-    with interpretation(lazy):
+    with lazy:
         unrolled_factors, unrolled_vars, remaining_plates = partial_unroll(
             factors, vars1 | vars2, plate_to_step
         )
@@ -1173,7 +1172,7 @@ def test_modified_partial_sum_product_8(
     )
     actual = reduce(prod_op, factors2)
 
-    with interpretation(lazy):
+    with lazy:
         unrolled_factors, unrolled_vars, remaining_plates = partial_unroll(
             factors, vars1 | vars2, plate_to_step
         )
@@ -1354,7 +1353,7 @@ def test_modified_partial_sum_product_9(
         "tones": {},
     }
 
-    with interpretation(lazy if use_lazy else eager):
+    with (lazy if use_lazy else eager):
         factors1 = modified_partial_sum_product(
             sum_op, prod_op, factors, vars1, plate_to_step
         )
@@ -1364,7 +1363,7 @@ def test_modified_partial_sum_product_9(
         actual = reduce(prod_op, factors2)
     actual = apply_optimizer(actual)
 
-    with interpretation(lazy):
+    with lazy:
         unrolled_factors, unrolled_vars, remaining_plates = partial_unroll(
             factors, vars1 | vars2, plate_to_step
         )
@@ -1534,7 +1533,7 @@ def test_modified_partial_sum_product_10(
     )
     actual = reduce(prod_op, factors2)
 
-    with interpretation(lazy):
+    with lazy:
         unrolled_factors, unrolled_vars, remaining_plates = partial_unroll(
             factors, vars1 | vars2, plate_to_step
         )
@@ -1749,7 +1748,7 @@ def test_modified_partial_sum_product_11(
         "tones": {},
     }
 
-    with interpretation(lazy if use_lazy else eager):
+    with (lazy if use_lazy else eager):
         factors1 = modified_partial_sum_product(
             sum_op, prod_op, factors, vars1, plate_to_step
         )
@@ -1759,7 +1758,7 @@ def test_modified_partial_sum_product_11(
         actual = reduce(prod_op, factors2)
     actual = apply_optimizer(actual)
 
-    with interpretation(lazy):
+    with lazy:
         unrolled_factors, unrolled_vars, remaining_plates = partial_unroll(
             factors, vars1 | vars2, plate_to_step
         )
@@ -2097,7 +2096,7 @@ def test_modified_partial_sum_product_13(
         "weeks": frozenset({("y_0", "y_prev", "y_curr")}),
     }
 
-    with interpretation(lazy if use_lazy else eager):
+    with (lazy if use_lazy else eager):
         factors1 = modified_partial_sum_product(
             sum_op, prod_op, factors, vars1, plate_to_step
         )
@@ -2107,7 +2106,7 @@ def test_modified_partial_sum_product_13(
         actual = reduce(prod_op, factors2)
     actual = apply_optimizer(actual)
 
-    with interpretation(lazy):
+    with lazy:
         unrolled_factors, unrolled_vars, remaining_plates = partial_unroll(
             factors, vars1 | vars2, plate_to_step
         )
@@ -2296,7 +2295,7 @@ def test_modified_partial_sum_product_14(
     )
     actual = reduce(prod_op, factors2)
 
-    with interpretation(lazy):
+    with lazy:
         unrolled_factors, unrolled_vars, remaining_plates = partial_unroll(
             factors, vars1 | vars2, plate_to_step
         )
@@ -2391,7 +2390,7 @@ def test_modified_partial_sum_product_16(
     )
     actual = reduce(prod_op, factors2)
 
-    with interpretation(lazy):
+    with lazy:
         unrolled_factors, unrolled_vars, remaining_plates = partial_unroll(
             factors, vars1 | vars2, plate_to_step
         )
@@ -2557,7 +2556,7 @@ def test_modified_partial_sum_product_17(
         "time": frozenset({("x_0", "x_prev", "x_curr")}),
     }
 
-    with interpretation(lazy if use_lazy else eager):
+    with (lazy if use_lazy else eager):
         factors1 = modified_partial_sum_product(
             sum_op, prod_op, factors, vars1, plate_to_step
         )
@@ -2567,7 +2566,7 @@ def test_modified_partial_sum_product_17(
         actual = reduce(prod_op, factors2)
     actual = apply_optimizer(actual)
 
-    with interpretation(lazy):
+    with lazy:
         unrolled_factors, unrolled_vars, remaining_plates = partial_unroll(
             factors, vars1 | vars2, plate_to_step
         )
@@ -2635,7 +2634,7 @@ def test_sequential_sum_product(
         for t in range(num_steps)
     )
     reduce_vars = frozenset("t_{}".format(t) for t in range(1, num_steps))
-    with interpretation(reflect):
+    with reflect:
         expected = sum_product(sum_op, prod_op, operands, reduce_vars)
     expected = apply_optimizer(expected)
     expected = expected(**{"t_0": "prev", "t_{}".format(num_steps): "curr"})
@@ -2690,7 +2689,7 @@ def test_sequential_sum_product_multi(
     time = Variable("time", Bint[num_steps])
     step = {"x_prev": "x_curr", "y_prev": "y_curr"}
 
-    with interpretation(moment_matching):
+    with moment_matching:
         actual = impl(sum_op, prod_op, trans, time, step)
         expected_inputs = batch_inputs.copy()
         expected_inputs.update(
@@ -2802,7 +2801,7 @@ def test_sequential_sum_product_bias_2(num_steps, num_sensors, dim):
         OrderedDict(time=Bint[num_steps]),
         dtype=2,
     )
-    with interpretation(eager_or_die):
+    with eager_or_die:
         factor = trans + obs(bias=bias[sensor_id]) + bias_dist
     assert set(factor.inputs) == {"time", "bias", "x_prev", "x_curr"}
